improve abstraction
